Refactor directory structure to domain-driven architecture
Organized all source files and headers into logically grouped subdirectories:
- `Characters/` (Player, Enemy, Character)
- `Core/` (Game, Extern, main)
- `Scene/` (Start, Home, DungeonMenu, Dungeon, Congratulations, GameOver)
- `System/` (AStar, LTexture)
- `GenerateDungeon/` (Generator, AreaDivide, RRA, Tile, Enum, Const)
- `UI/` (Log, Button)
- `Utils/` (Util, Color, RectUtils, Texture, Ivec2)

Updated all `#include` paths throughout the codebase, refactored the Makefile to compile from these new directories and resolved various relative path issues.
