+++ src/Core/Extern.cpp
