+++ src/Scene/6_GameOver.h
@@ -5,12 +5,12 @@
 #include <SDL_ttf.h>
 #include <SDL_mixer.h>
 
-#include "../UI/Button.h"
-#include "../Functions/Color.h"
-#include "../GenerateDungeon/Const.h"
-#include "../GenerateDungeon/Enum.h"
+#include "UI/Button.h"
+#include "Utils/Color.h"
+#include "GenerateDungeon/Const.h"
+#include "GenerateDungeon/Enum.h"
 
-#include "../Game.h"
+#include "Core/Game.h"
 
 class GameOver
 {
