+++ src/UI/Log.cpp
