+++ src/Utils/Util.h
@@ -5,13 +5,14 @@
 
 inline std::string resource_path(const char* relative_path) {
     const char* appdir = getenv("APPDIR");
+    std::string full_path;
     
     if (appdir) {
         // AppImage内で実行されている場合
-        return std::string(appdir) + "/usr/share/rogue/assets/" + relative_path;
+        full_path = std::string(appdir) + "/usr/share/rogue/assets/" + relative_path;
     } else {
         // 通常の実行の場合
-        return std::string("assets/") + relative_path;
+        full_path = std::string("assets/") + relative_path;
     }
     
     return full_path;
