--- src/main.cpp
@@ -1,15 +1,0 @@
-#define __DEBUG_
-
-#include "Game.h"
-
-int main(int argc, char **argv)
-{
-    Game game;
-    bool success = game.Init();
-    if(success)
-    {
-        game.RunLoop();
-    }
-    game.Shutdown();
-    return 0;
-}
